api/requirements.txt
fastapi
uvicorn
pandas
requests
faiss-cpu
together
sentence-transformers
python-multipart
openpyxl
ruff
pre-commit
